requirements.txt
Django==3.1.4
PyMySQL==0.10.1
djangorestframework
openpyxl==3.0.5
django-cors-headers